// apps/backend/src/enums/status.ts
export enum StatusBooking {
    PENDING = 'PENDING',
    SUCCESS = 'SUCCESS',
    FAILED = 'FAILED',
    CANCELLED = 'CANCELLED'
}

export enum StatusSeat {
    HELD = 'HELD',
    CONFIRMED = 'CONFIRMED',
    RELEASED = 'RELEASED'
}


export enum StatusPayment {
    PENDING = 'PENDING',
    SUCCESS = 'SUCCESS',
    FAILED = 'FAILED',
    CANCELLED = 'CANCELLED'
}

export enum PaymentBank {
    MOMO = 'MOMO',
    STRIPE = 'STRIPE'
}